refactor: 요금 도메인 패키지 분리
