+++ src/main/java/nextstep/path/application/PathService.java
@@ -2,11 +2,10 @@ package nextstep.path.application;
 
 import nextstep.line.domain.SectionRepository;
 import nextstep.path.application.dto.PathsResponse;
-import nextstep.path.domain.Fare;
+import nextstep.fare.domain.Fare;
 import nextstep.path.domain.PathType;
 import nextstep.path.domain.ShortestPath;
 import nextstep.path.ui.exception.SameSourceAndTargetException;
-import nextstep.station.application.StationService;
 import nextstep.station.application.dto.StationResponse;
 import nextstep.station.domain.Station;
 import org.springframework.stereotype.Service;
